// src/app/domain/models/user-internal/create-user-internal.ts
export interface CreateUserInternal {
  readonly username: string;
  readonly password: string;
  readonly email: string;
  readonly role: 'INTERNAL' | 'ADMIN';
  readonly firstName: string;
  readonly paternalLastName: string;
  readonly maternalLastName: string;
  readonly phoneNumber: string;
  readonly dni: string;
  readonly permissions: readonly string[];
}